+++ x/emissions/module/module.go
@@ -18,7 +18,7 @@ import (
 	"github.com/allora-network/allora-chain/x/emissions/keeper/queryserver"
 	migrationV10 "github.com/allora-network/allora-chain/x/emissions/migrations/v10"
 	migrationV11 "github.com/allora-network/allora-chain/x/emissions/migrations/v11"
-	migrationV14 "github.com/allora-network/allora-chain/x/emissions/migrations/v14"
+	migrationV13 "github.com/allora-network/allora-chain/x/emissions/migrations/v13"
 	migrationV2 "github.com/allora-network/allora-chain/x/emissions/migrations/v2"
 	migrationV3 "github.com/allora-network/allora-chain/x/emissions/migrations/v3"
 	migrationV4 "github.com/allora-network/allora-chain/x/emissions/migrations/v4"
@@ -45,7 +45,7 @@ var (
 )
 
 // ConsensusVersion defines the current module consensus version.
-const ConsensusVersion = 14
+const ConsensusVersion = 13
 
 type AppModule struct {
 	cdc    codec.Codec
@@ -157,14 +157,9 @@ func (am AppModule) RegisterServices(cfg module.Configurator) {
 		panic(fmt.Sprintf("failed to migrate x/%s from version 11 to 12: %v", types.ModuleName, err))
 	}
 	if err := cfg.RegisterMigration(types.ModuleName, 12, func(ctx sdk.Context) error {
-		return nil
+		return migrationV13.MigrateStore(ctx, am.keeper)
 	}); err != nil {
 		panic(fmt.Sprintf("failed to migrate x/%s from version 12 to 13: %v", types.ModuleName, err))
-	}
-	if err := cfg.RegisterMigration(types.ModuleName, 13, func(ctx sdk.Context) error {
-		return migrationV14.MigrateStore(ctx, am.keeper)
-	}); err != nil {
-		panic(fmt.Sprintf("failed to migrate x/%s from version 13 to 14: %v", types.ModuleName, err))
 	}
 }