+++ CHANGELOG.md
@@ -70,7 +70,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [#851](https://github.com/allora-network/allora-chain/pull/851) Enforce recommended settings
 * [#835](https://github.com/allora-network/allora-chain/pull/835): Introduce the `x/scheduler` module to manage tasks scheduling
 * [#831](https://github.com/allora-network/allora-chain/pull/831) Allow topic creators to update topic configs
-* [#893](https://github.com/allora-network/allora-chain/pull/893) Make `c_norm` editable per topic and add migration v14
+* [#893](https://github.com/allora-network/allora-chain/pull/893) Make `c_norm` editable per topic and add migration v13
 
 ### Changed